requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
python-decouple
django-filter
drf_yasg
celery
redis
redis-server
redis-tools
